Two classes for adapted RDP3D algorithm are added
The 3d version of RDP algorithm RamerDuglasPeuckerAlgorithm3D.cs isn't work properly. So I wrote 2d version for rows and columns of the points. Each point is an ProcessedPoint.cs object and has rowStatus and columnStatus. If point has both positive statuses, it will be a part of optimized map.